Added documentation

+++ src/ros/ublox_ros.cpp
@@ -82,7 +82,7 @@ UBLOX_ROS::UBLOX_ROS() :
         uint16_t* rover_port = new uint16_t[std::max(1, rover_quantity)];
         
         //Account for the case when no numbers are used for the first rover.
-        int j = 0;
+        uint8_t j = 0;
         if(nh_private_.hasParam("local_host")) {
             //The first local host corresponds to the first rover.
             local_host[0] = nh_private_.param<std::string>("local_host", "localhost");
@@ -159,12 +159,12 @@ UBLOX_ROS::UBLOX_ROS() :
 
         // Get Constallation settings
         uint32_t constellation [6];
-        int gps = nh_.param<int>("GPS", 1);
-        int glonas = nh_.param<int>("GLONAS", 0);
-        int beidou = nh_.param<int>("BEIDOU", 0);
-        int galileo = nh_.param<int>("GALILEO", 1);
+        uint8_t gps = nh_.param<uint8_t>("GPS", 1);
+        uint8_t glonas = nh_.param<uint8_t>("GLONAS", 0);
+        uint8_t beidou = nh_.param<uint8_t>("BEIDOU", 0);
+        uint8_t galileo = nh_.param<uint8_t>("GALILEO", 1);
 
-        int j = 0;
+        uint8_t j = 0;
         if(nh_private_.hasParam("local_host")) {
 
             local_host[0] = nh_private_.param<std::string>("local_host", "localhost");
@@ -229,7 +229,6 @@ UBLOX_ROS::UBLOX_ROS() :
     {
         ublox_->initLogFile(log_filename);
         //ublox_->readFile(log_filename);
-
     }
 }
 
@@ -465,7 +464,7 @@ void UBLOX_ROS::obsCB(const ublox::UBX_message_t &ubx_msg, uint8_t f9pID)
 //        case ublox::Observation::BEIDOU_B2I_D2:
 //            out.obs[i].freq = Ephemeris::BEIDOU_FREQ_B2;
 //            break;
-        default:      // and tip of the arrow for /RelPos
+        default:
             out.obs[i].freq = 0;
             break;
         }
